OPHJOD-1255: Update VirtualAssistant layout
